% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dtitk.R
\name{download_dtitk}
\alias{download_dtitk}
\title{Download DTI-TK}
\usage{
download_dtitk(os = c("osx", "linux"))
}
\arguments{
\item{os}{Operating system to download}
}
\description{
Download DTI-TK tarball from NITRC website
}

