% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_dtitk.R
\name{install_dtitk}
\alias{install_dtitk}
\title{Install DTI-TK}
\usage{
install_dtitk(tarfile, verbose = TRUE)
}
\arguments{
\item{tarfile}{Tarball file (with extension \code{.tar.gz})}

\item{verbose}{Print output of \code{untar}}
}
\value{
Logical indicating files were there
}
\description{
Install DTI-TK from the tarball you downloaded
}

